pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yandex-auto-up"
dynamic = ["version"]
description = "Low-resource Yandex Cloud VM auto-start watchdog with a polished CLI panel."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "censorny" }
]
dependencies = [
  "httpx>=0.27,<1",
  "pydantic>=2.7,<3",
  "PyJWT[crypto]>=2.8,<3",
  "rich>=13.7,<14",
  "typer>=0.12,<1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9"
]

[project.scripts]
yauto = "yauto.cli.app:main"
yauto-daemon = "yauto.daemon.main:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.package-data]
yauto = ["version.json"]

[tool.setuptools.dynamic]
version = { attr = "yauto.__version__" }

[tool.setuptools.packages.find]
where = ["src"]
